This version read the *.xlsx files corresponsing to individual pixels extracted from PLANET timeseries for USFS flights-summer 2025, priorities areas 2 and 3.
The *.xlsx files contains weekly NDVI, RED and NIR bands for weeks 1 to 40 of 2025.

This script create new *.csv files containing the File, NDVI, year and day fields and sorted by year and day.

It calculates multiple change points and their corresponding max_posterio_probability and later the highest_max_posterior_probability is calculated.

At the end of the script, the results are plotted as frequency histogram and a cumilative frequency plot.

Results are saved in an R-dataframe and a *.csv file.